# laravel-synoptic-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger API Documentation",
        "description": "A simple API for display synoptic data",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Roman Mah"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Laravel API server"
        }
    ],
    "paths": {
        "/weather": {
            "get": {
                "summary": "Get a data about current weather",
                "description": "Get a data about current weather",
                "operationId": "weatherData",
                "parameters": [
                    {
                        "name": "cities",
                        "in": "query",
                        "description": "Cities",
                        "required": false,
                        "schema": {
                            "schema": "ArrayField",
                            "properties": {
                                "city[]": {
                                    "type": "array",
                                    "items": {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            }
                                        ]
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Weather",
            "description": "Weather data"
        }
    ],
    "security": [
        []
    ]
}